Splits cleanShorthandValues out and adds tests
